refact: email버전,phone 버전 회원가입으로 변경 및 hooks migration
