+++ frontend/src/@types/user.ts
@@ -3,8 +3,19 @@ type RequestLoginUserType = {
   password: string;
 };
 
-type RequestSignupUserType = {
+type RequestEmailSignupUserType = {
   email: string;
+  password: string;
+  name: string;
+  birth: {
+    year: number;
+    month: number;
+    day: number;
+  };
+};
+
+type RequestPhoneSignupUserType = {
+  phone: string;
   password: string;
   name: string;
   birth: {
